+++ src/ReindexerNet.Core/Model/JoinedDef.cs
@@ -49,7 +49,7 @@ namespace ReindexerNet {
     /// <value>Maximum count of returned items</value>
     [DataMember(Name="limit", EmitDefaultValue=false)]
     [JsonPropertyName("limit")]
-    public int? Limit { get; set; }
+    public long? Limit { get; set; }
 
     /// <summary>
     /// Offset of first returned item
@@ -57,7 +57,7 @@ namespace ReindexerNet {
     /// <value>Offset of first returned item</value>
     [DataMember(Name="offset", EmitDefaultValue=false)]
     [JsonPropertyName("offset")]
-    public int? Offset { get; set; }
+    public long? Offset { get; set; }
 
     /// <summary>
     /// Join ON statement
